Update 1.6.5
Updater Fix:

In this version (1.6.5) i fixed the Updater Issue and now, the data fetch is every 7 minutes and not 1 minute.

Bug Fix:

Fixed Minecraft Online Session Issue, in 1.6.4 and previous, when a player try to login with 2 connected accounts, it generate an exception, now the exception will be not thrown but the connection will be blocked.

New events:

PremiumSwitchEvent - When a player enable / disable PremiumLogin using /premium for itself. (SwitchType -> getType() will be or PREMIUMLOGINENABLED or PREMIUMLOGINDISABLED).

PremiumStaffSwitchEvent - Similar to the PremiumSwitchEvent, but, this, will be fired when is a staff player to enable / disable PremiumLogin to a player. (/premiumadd, /premiumremove)

Config Fix 2:

Fixed the configuration language support issue.
Now, you can use the spanish ñ or other chars like "è" "ò"::

BUT, Note: You need to specify them using codes.yml.
Example, ñ is .N., you will see into codes.yml how to type this on the config.

ONE-TIME REQ:

YOU NEED TO SAVE YOUR MESSAGES, THEN DELETE THE CONFIG.YML AND RESTART THE SERVER. (It will generate the new 1.6.5 Configuration.)